requirements.txt
pandas
pandera
PyYAML
requests
eth-abi
eth-utils
pycryptodome
psycopg2-binary
python-dotenv
matplotlib
plotly
seaborn
